+++ R/eucast_rules.R
@@ -163,6 +163,7 @@ EUCAST_VERSION_EXPERT_RULES <- "3.1, 2016"
 #'   }
 #' @inheritSection AMR Read more on our website!
 #' @examples
+#' \donttest{
 #' a <- data.frame(mo = c("Staphylococcus aureus",
 #'                        "Enterococcus faecalis",
 #'                        "Escherichia coli",
@@ -198,7 +199,6 @@ EUCAST_VERSION_EXPERT_RULES <- "3.1, 2016"
 #' # 5 Pseudomonas aeruginosa    R    R    -    -    R    R    R
 #'
 #'
-#' \donttest{
 #' # do not apply EUCAST rules, but rather get a data.frame
 #' # with 18 rows, containing all details about the transformations:
 #' c <- eucast_rules(a, verbose = TRUE)