+++ git_premaster.sh
@@ -37,7 +37,7 @@ sed -i -- "s/^Version: .*/Version: ${new_version}/" DESCRIPTION
 # update 1st line of NEWS.md
 sed -i -- "1s/.*/# AMR ${new_version}/" NEWS.md
 # add date to 2nd line of NEWS.md
-sed -i -- "2s/.*/\<small\>Last updated: $(date '+%d-%b-%Y')\<\/small\>/" NEWS.md
+sed -i -- "2s/.*/## \<small\>Last updated: $(date '+%d-%b-%Y')\<\/small\>/" NEWS.md
 rm *-- || true
 echo "• First 3 lines of DESCRIPTION:"
 head -3 DESCRIPTION